Page adder, setup script and bug fixes
There are a lot more files now:
FB_page_adder_bot.py is a support script to the main one that makes it possible to add FB pages to the bot AND Telegram channel via Telegram itself, both for the admin and for users. Every request is to be manually verified by the admin.
FB_bot_setup.py is a setup script for everyone who would like to setup their own Facebook bot on their own PC/RaspberryPi/whatever. It's needed because now a lot of the settings of the bot are given via .ini files, which i can't upload as they keep the Bot's Token and other goodies, but are generated by the setup script, just give him the data it asks for.
The setup is not complete, it's very much in alpha and is VERY unpolished, but should work.
There were modifications to FB_bot.py itself too, mostly bug fixes, the functions needed tpo make in interface with the other scripts, shifting to .ini files, and more little things
Another change is in the .csv file, it now doesn't need the TOKEN data at [4], that's now loaded via .ini file, now at [4] there should be the channel ID